refactor(services): extract shared buildAllServicesLayer helper; normalize apex-replay-debugger
Review suggestions for #7267:

1. Extract the duplicated buildAllServicesLayer + ExtensionProviderServiceLive
   into @salesforce/effect-ext-utils. Each consumer's services/extensionProvider.ts
   becomes a thin wrapper that supplies its channel-name fallback. Eliminates the
   ~270-token clones jscpd was flagging.

2. Normalize apex-replay-debugger to the typed AllServicesLayer pattern used by
   apex, apex-debugger, apex-oas, core, and lwc. Drops the allServicesLayerRef.ts
   opaque-cast indirection.

The opaque cast was masking a real type gap: prebuiltServicesDependencies'
runtime context (built from globalLayers in the services extension) includes
TraceFlagService and ExecuteAnonymousService, but the static type union under-
reported them. Apex-replay-debugger consumes both. Adding them to the
SalesforceVSCodeServicesApi prebuiltServicesDependencies type union (additive,
non-breaking) makes the typed pattern usable.

Lightning's services dir is pre-existing (PR #7260) and uses the same opaque
indirection - left unchanged here; can be normalized in a follow-up.

Made-with: Cursor
