+++ packages/effect-ext-utils/src/index.ts
@@ -12,6 +12,7 @@ export {
   getServicesApi
 } from './extensionProvider';
 
+export { buildAllServicesLayer } from './allServicesLayer';
 export { ExtensionPackageJsonSchema, type ExtensionPackageJson } from './extensionPackageJson';
 export { closeExtensionScope, getExtensionScope } from './extensionScope';
 export type { SalesforceVSCodeServicesApi } from './extensionProvider';
