+++ packages/salesforcedx-vscode-apex/src/services/extensionProvider.ts
@@ -5,43 +5,10 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import {
-  ExtensionPackageJsonSchema,
-  ExtensionProviderService,
-  type ExtensionPackageJson,
-  getServicesApi
-} from '@salesforce/effect-ext-utils';
-import * as Effect from 'effect/Effect';
-import * as Layer from 'effect/Layer';
-import * as Schema from 'effect/Schema';
+import { buildAllServicesLayer as buildAllServicesLayerCore } from '@salesforce/effect-ext-utils';
 import type { ExtensionContext } from 'vscode';
 
-const ExtensionProviderServiceLive = Layer.effect(
-  ExtensionProviderService,
-  Effect.sync(() => ({ getServicesApi }))
-);
-
-export const buildAllServicesLayer = (context: ExtensionContext) =>
-  Layer.unwrapEffect(
-    Effect.gen(function* () {
-      const extensionProvider = yield* ExtensionProviderService;
-      const api = yield* extensionProvider.getServicesApi;
-      const emptyPjson: ExtensionPackageJson = {};
-      const pjson = yield* Schema.decodeUnknown(ExtensionPackageJsonSchema)(context.extension.packageJSON).pipe(
-        Effect.catchAll(() => Effect.succeed(emptyPjson))
-      );
-      const channelLayer = api.services.ChannelServiceLayer(pjson.displayName ?? 'Apex');
-      const errorHandlerWithChannel = Layer.provide(api.services.ErrorHandlerService.Default, channelLayer);
-      return Layer.mergeAll(
-        Layer.succeedContext(api.services.prebuiltServicesDependencies),
-        ExtensionProviderServiceLive,
-        api.services.ExtensionContextServiceLayer(context),
-        api.services.SdkLayerFor(context),
-        channelLayer,
-        errorHandlerWithChannel
-      );
-    }).pipe(Effect.provide(ExtensionProviderServiceLive))
-  );
+export const buildAllServicesLayer = (context: ExtensionContext) => buildAllServicesLayerCore(context, 'Apex');
 
 export let AllServicesLayer: ReturnType<typeof buildAllServicesLayer>;
