+++ packages/salesforcedx-vscode-core/src/services/extensionProvider.ts
@@ -5,46 +5,19 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { ExtensionProviderService, getServicesApi } from '@salesforce/effect-ext-utils';
-import * as Effect from 'effect/Effect';
-import * as Layer from 'effect/Layer';
+import { buildAllServicesLayer as buildAllServicesLayerCore } from '@salesforce/effect-ext-utils';
 import type { ExtensionContext } from 'vscode';
-
-const ExtensionProviderServiceLive = Layer.effect(
-  ExtensionProviderService,
-  Effect.sync(() => ({
-    getServicesApi
-  }))
-);
 
 /**
  * Factory for a Layer that provides all services from the SalesforceVSCodeServicesApi.
  * Pass the ExtensionContext to include a working ExtensionContextServiceLayer.
- * When context is not provided, ExtensionContextService.Default is used (fails if getContext is called).
  */
 export const buildAllServicesLayer = (context: ExtensionContext) =>
-  Layer.unwrapEffect(
-    Effect.gen(function* () {
-      const extensionProvider = yield* ExtensionProviderService;
-      const api = yield* extensionProvider.getServicesApi;
-      const displayName = context.extension.packageJSON?.displayName ?? 'Salesforce CLI';
-      const channelLayer = api.services.ChannelServiceLayer(displayName);
-      const errorHandlerWithChannel = Layer.provide(api.services.ErrorHandlerService.Default, channelLayer);
-      return Layer.mergeAll(
-        Layer.succeedContext(api.services.prebuiltServicesDependencies),
-        ExtensionProviderServiceLive,
-        errorHandlerWithChannel,
-        api.services.ExtensionContextServiceLayer(context),
-        api.services.SdkLayerFor(context),
-        channelLayer
-      );
-    }).pipe(Effect.provide(ExtensionProviderServiceLive))
-  );
+  buildAllServicesLayerCore(context, 'Salesforce CLI');
 
 /**
  * Layer that provides all services from the SalesforceVSCodeServicesApi.
- * Uses ExtensionContextService.Default (fails if getContext is called).
- * Use buildAllServicesLayer(context) to provide a working ExtensionContextService.
+ * Set via setAllServicesLayer during extension activation; consumed by getRuntime().
  */
 export let AllServicesLayer: ReturnType<typeof buildAllServicesLayer>;
